📦 Rename: 지원현황 모달 이름 및 위치 변경
- ViewStatus에서 사용되는 모달 Modal 폴더로 이동 및 이름 변경
- 모달 위치 변경에 따른 import 위치 수정

Ref: #38
