+++ src/components/Modal/Modal.css
@@ -28,7 +28,7 @@
     height: 9rem;
     border-radius: 5rem;
     background-color: #2ab22a;
-  };
+  }
   .IconCircleRed {
     display: flex;
     justify-content: center;
@@ -36,7 +36,7 @@
     width: 9rem;
     height: 9rem;
     border-radius: 5rem;
-    background-color: #D62117;
+    background-color: #d62117;
   }
 }
 
@@ -91,7 +91,7 @@
   font-size: 1.4rem;
   font-weight: 300;
   line-height: 2rem; /* 142.857% */
-  &:hover{
+  &:hover {
     opacity: 0.8;
   }
 }
@@ -112,7 +112,7 @@
   font-size: 1.4rem;
   font-weight: 300;
   line-height: 2rem; /* 142.857% */
-  &:hover{
+  &:hover {
     opacity: 0.8;
   }
 }
