+++ src/components/Modal/index.js
@@ -1,0 +1,4 @@
+import EmailRequestModal from "./EmailRequest";
+import SignupComplete from "./SignupComplete";
+
+export { EmailRequestModal, SignupComplete };
